+++ R/LDA.R
@@ -63,9 +63,13 @@ LDA_set <- function(document_term_table, topics = 2, nseeds = 1,
 
 #' @title Calculate the log likelihood of a VEM LDA model fit
 #'
-#' @description Imported calculations from topicmodels package, as applied to
-#'   Latent Dirichlet Allocation fit with Variational Expectation Maximization
-#'   via \code{\link[topicmodels]{LDA}}. 
+#' @description Imported but updated calculations from topicmodels package, as
+#'   applied to Latent Dirichlet Allocation fit with Variational Expectation 
+#'   Maximization via \code{\link[topicmodels]{LDA}}. 
+#'
+#' @details The number of degrees of freedom is 1 (for alpha) plus the number
+#'   of entries in the document-topic matrix. The number of observations is 
+#'   the number of entries in the document-term matrix.
 #'
 #' @param object A \code{LDA_VEM}-class object.
 #'
@@ -75,9 +79,18 @@ LDA_set <- function(document_term_table, topics = 2, nseeds = 1,
 #'   (degrees of freedom) and \code{nobs} (number of observations) values.
 #'
 #' @references 
+#'   Buntine, W. 2002. Variational extentions to EM and multinomial PCA. 
+#'   \emph{European Conference on Machine Learning, Lecture Notes in Computer 
+#'   Science} \strong{2430}:23-34. \href{https://bit.ly/327sltH}{link}.
+#'
 #'   Grun B. and K. Hornik. 2011. topicmodels: An R Package for Fitting Topic
 #'   Models. \emph{Journal of Statistical Software} \strong{40}:13.
 #'   \href{https://www.jstatsoft.org/article/view/v040i13}{link}.
+#'
+#'   Hoffman, M. D., D. M. Blei, and F. Bach. 2010. Online learning for 
+#'   latent Dirichlet allocation. \emph{Advances in Neural Information 
+#'   Processing Systems} \strong{23}:856-864.
+#'   \href{https://bit.ly/2LEr5sb}{link}.
 #'
 #' @export
 #'
@@ -85,7 +98,7 @@ logLik.LDA_VEM <- function(object, ...){
   val <- sum(object@loglikelihood)
   df <- as.integer(object@control@estimate.alpha) + length(object@beta)
   attr(val, "df") <- df
-  attr(val, "nobs") <- object@Dim[1]
+  attr(val, "nobs") <- object@Dim[1] * object@Dim[2]
   class(val) <- "logLik"
   val
 }